+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 See Git checking messages for full history.
 
 ## 10.1.1.dev0 (2025-xx-xx)
-- 
+- Linux: check the server for Xrandr support version (#417)
 - :heart: contributors: @
 
 ## 10.1.0 (2025-08-16)