+++ CHANGELOG.md
@@ -3,8 +3,8 @@
 See Git checking messages for full history.
 
 ## 10.1.1.dev0 (2025-xx-xx)
-- 
-- :heart: contributors: @
+- Linux: check the server for Xrandr support version (#417)
+- :heart: contributors: @jholveck
 
 ## 10.1.0 (2025-08-16)
 - Mac: up to 60% performances improvement by taking screenshots at nominal resolution (e.g. scaling is off by default). To enable back scaling, set `mss.darwin.IMAGE_OPTIONS = 0`. (#257)