+++ CHANGELOG.md
@@ -4,6 +4,7 @@ See Git checking messages for full history.
 
 ## 10.1.1.dev0 (2025-xx-xx)
 - Linux: check the server for Xrandr support version (#417)
+- Linux: improve typing and error messages for X libraries (#418)
 - :heart: contributors: @jholveck
 
 ## 10.1.0 (2025-08-16)